Added test for finalizer.
